requirements.txt
requests==2.31.0
Jinja2==3.1.2
python-dotenv==1.0.0
PyYAML==6.0.1
